Run checks for various LiteX packages
This commit enables the checkPhase for various LiteX packages and adds
the checkInput dependencies which are required for the packages.

Because the various LiteX packages depend on each other in the check
phase, this is not trivial to resolve. One needs to build some
packages unchecked, use that as a check input for other packages, such
that these packages can be used to build a checked version of the
target package.

This commit uses a generic function to abstract this complexity and
uses fixed-point evaluation to build both a set of unchecked and,
using these to break recursion, checked packages.

As an example, take the packages `litex` and `liteeth`. `liteeth`
requires `litex` as a proper build input. However, `litex` uses
`liteeth` to run a set of integration tests. To break the recursion,
we build a `liteeth-unchecked` to build `litex` and use that to build
a properly checked `liteeth`. Then, in one final round, all packages
are built with only checked dependencies as inputs. Thus every package
and every dependecy of every package is a properly tested package.

Not all tests of every package work yet (perhaps because of missing
check  dependencies), but generally the concept has proven itself to
be quite useful.

Signed-off-by: Leon Schuermann <leon@is.currently.online>